+++ tests/test_integration.py
@@ -146,6 +146,10 @@ def test_repo_total_untestables(test_repo):
     assert test_repo.total_untestables() > 0
 
 
+def test_repo_total_size(test_repo):
+    assert test_repo.total_size() > 0
+
+
 def test_repo_languages_by_letter(test_repo):
     assert len(test_repo.languages_by_letter("p")) > 0
 
@@ -195,6 +199,33 @@ def test_language_doc_modified(test_repo):
     assert language.doc_modified() is not None
 
 
+@pytest.mark.parametrize(
+    "language,expected_color",
+    [
+        # Not in Linguist
+        ("Algol68", subete.repo.OTHER_LANGUAGE_COLOR),
+        # In Linguist
+        ("C", "#555555"),
+        ("C++", "#F34B7D"),
+        ("C#", "#178600"),
+        (r"F\*", "#572E30"),
+        ("Python", "#3572A5"),
+        # Additional colors
+        (r"C\*", "#6725A3"),
+        ("Commodore Basic", "#141AC9"),
+        ("Dale", "#8A9A5B"),
+    ]
+)
+def test_language_color(language, expected_color, test_repo):
+    language: subete.LanguageCollection = test_repo[language]
+    assert language.color() == expected_color
+
+
+def test_language_percentage(test_repo):
+    language: subete.LanguageCollection = test_repo["Python"]
+    assert 0.0 < language.percentage() < 100.0
+
+
 def test_project_doc_author(test_repo):
     project: subete.Project = test_repo["Python"]["Selection Sort"].project()
     assert "Parker Johansen" in project.doc_authors()
@@ -214,7 +245,9 @@ def test_project_doc_modified(test_repo):
 def test_repo():
     with patch("subete.repo.tempfile.TemporaryDirectory") as mock:
         mock.side_effect = [SAMPLE_PROGRAMS_TEMP_DIR, SAMPLE_PROGRAMS_WEBSITE_TEMP_DIR]
-        yield subete.load()
+        repo = subete.load()
+        repo.set_additional_language_colors("tests/additional-language-colors.yml")
+        yield repo
 
     SAMPLE_PROGRAMS_TEMP_DIR.cleanup()
     SAMPLE_PROGRAMS_WEBSITE_TEMP_DIR.cleanup()